code cleanup and commenting brought about when documenting for jhu report
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder@886 1f5c12ca-751b-0410-a591-d2e778427230

+++ trunk/moses/src/LanguageModel.cpp
@@ -41,7 +41,7 @@ LanguageModel::LanguageModel(bool registerScore)
 LanguageModel::~LanguageModel() {}
 
 // don't inline virtual funcs...
-unsigned int LanguageModel::GetNumScoreComponents() const
+size_t LanguageModel::GetNumScoreComponents() const
 {
 	return 1;
 }
